Made Hours a double

+++ JohnsonTaxes.java
@@ -3,14 +3,14 @@ public class JohnsonTaxes {
   public static void main(String[] args) {
     System.out.println("Enter the hours worked in a week: ");
     Scanner keyboard = new Scanner(System.in);
-    int hours = keyboard.nextInt();
+    double hours = keyboard.nextDouble();
     System.out.println("Enter the hourly pay rate: ");
     double payRate = keyboard.nextDouble();
     final double FED_TAX = .18;
     final double FICA = .08;
     final double STATE_TAX = .032;
     final double LOCAL_TAX = .02;
-    System.out.printf("\nHours worked in a week: %23d\nHourly pay rate: \t\t\t\t\t$%10.2f\n\n", hours,payRate);
+    System.out.printf("\nHours worked in a week: %23.2f\nHourly pay rate: \t\t\t\t\t$%10.2f\n\n", hours,payRate);
     double grossPay = hours*payRate;
     double fedRemoved = FED_TAX*grossPay;
     double FICARemoved = FICA*grossPay;
